pyproject.toml
[project]
name = "graphrag-weaviate-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "attrs==25.4.0",
    "authlib==1.6.5",
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "cryptography==46.0.3",
    "dataclasses-json==0.6.7",
    "deprecation==2.1.0",
    "frozenlist==1.8.0",
    "fsspec==2024.12.0",
    "greenlet==3.2.4",
    "grpcio==1.76.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "idna==3.11",
    "json-repair==0.44.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==1.0.3",
    "langchain-classic==1.0.0",
    "langchain-community==0.4.1",
    "langchain-core==1.0.3",
    "langchain-neo4j==0.6.0",
    "langchain-ollama==1.0.0",
    "langchain-text-splitters==1.0.0",
    "langchain-weaviate==0.0.6",
    "langgraph==1.0.2",
    "langgraph-checkpoint==3.0.1",
    "langgraph-prebuilt==1.0.2",
    "langgraph-sdk==0.2.9",
    "langsmith==0.4.41",
    "marshmallow==3.26.1",
    "multidict==6.7.0",
    "mypy-extensions==1.1.0",
    "neo4j==5.28.2",
    "neo4j-graphrag==1.10.1",
    "numpy==2.3.4",
    "ollama==0.6.0",
    "orjson==3.11.4",
    "ormsgpack==1.12.0",
    "packaging==25.0",
    "propcache==0.4.1",
    "protobuf==6.33.0",
    "pycparser==2.23",
    "pydantic==2.12.4",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.11.0",
    "pypdf==6.1.3",
    "python-dotenv==1.2.1",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "scipy==1.16.3",
    "simsimd==6.5.3",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.44",
    "tenacity==9.1.2",
    "types-pyyaml==6.0.12.20250915",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "urllib3==2.5.0",
    "validators==0.35.0",
    "weaviate-client==4.17.0",
    "xxhash==3.6.0",
    "yarl==1.22.0",
    "zstandard==0.25.0",
]
